Remove deprecated TextChannel.active_threads for Guild version
This also fills in the ThreadMember data from the endpoint
