+++ discord/guild.py
@@ -73,7 +73,7 @@ from .asset import Asset
 from .flags import SystemChannelFlags
 from .integrations import Integration, _integration_factory
 from .stage_instance import StageInstance
-from .threads import Thread
+from .threads import Thread, ThreadMember
 from .sticker import GuildSticker
 from .file import File
 
@@ -423,7 +423,9 @@ class Guild(Hashable):
 
         self.mfa_level: MFALevel = guild.get('mfa_level')
         self.emojis: Tuple[Emoji, ...] = tuple(map(lambda d: state.store_emoji(self, d), guild.get('emojis', [])))
-        self.stickers: Tuple[GuildSticker, ...] = tuple(map(lambda d: state.store_sticker(self, d), guild.get('stickers', [])))
+        self.stickers: Tuple[GuildSticker, ...] = tuple(
+            map(lambda d: state.store_sticker(self, d), guild.get('stickers', []))
+        )
         self.features: List[GuildFeature] = guild.get('features', [])
         self._splash: Optional[str] = guild.get('splash')
         self._system_channel_id: Optional[int] = utils._get_as_snowflake(guild, 'system_channel_id')
@@ -627,7 +629,6 @@ class Guild(Hashable):
             The returned channel or thread or ``None`` if not found.
         """
         return self._channels.get(channel_id) or self._threads.get(channel_id)
-
 
     def get_channel(self, channel_id: int, /) -> Optional[GuildChannel]:
         """Returns a channel with the given ID.
@@ -1590,6 +1591,35 @@ class Guild(Hashable):
             return channel
 
         return [convert(d) for d in data]
+
+    async def active_threads(self) -> List[Thread]:
+        """|coro|
+
+        Returns a list of active :class:`Thread` that the client can access.
+
+        This includes both private and public threads.
+
+        .. versionadded:: 2.0
+
+        Raises
+        ------
+        HTTPException
+            The request to get the active threads failed.
+
+        Returns
+        --------
+        List[:class:`Thread`]
+            The active threads
+        """
+        data = await self._state.http.get_active_threads(self.id)
+        threads = [Thread(guild=self, state=self._state, data=d) for d in data.get('threads', [])]
+        thread_lookup: Dict[int, Thread] = {thread.id: thread for thread in threads}
+        for member in data.get('members', []):
+            thread = thread_lookup.get(int(member['id']))
+            if thread is not None:
+                thread._add_member(ThreadMember(parent=thread, data=member))
+
+        return threads
 
     # TODO: Remove Optional typing here when async iterators are refactored
     def fetch_members(self, *, limit: int = 1000, after: Optional[SnowflakeTime] = None) -> MemberIterator:
