+++ discord/http.py
@@ -958,8 +958,8 @@ class HTTPClient:
         params['limit'] = limit
         return self.request(route, params=params)
 
-    def get_active_threads(self, channel_id: Snowflake) -> Response[threads.ThreadPaginationPayload]:
-        route = Route('GET', '/channels/{channel_id}/threads/active', channel_id=channel_id)
+    def get_active_threads(self, guild_id: Snowflake) -> Response[threads.ThreadPaginationPayload]:
+        route = Route('GET', '/guilds/{guild_id}/threads/active', guild_id=guild_id)
         return self.request(route)
 
     def get_thread_members(self, channel_id: Snowflake) -> Response[List[threads.ThreadMember]]:
